+++ app/src/main/java/org/breezyweather/common/basic/models/weather/Current.kt
@@ -36,6 +36,9 @@ data class Current(
     val airQuality: AirQuality? = null,
     val relativeHumidity: Float? = null,
     val dewPoint: Float? = null,
+    /**
+     * Pressure at sea level
+     */
     val pressure: Float? = null,
     val cloudCover: Int? = null,
     val visibility: Float? = null,